requirements.txt
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
numpy==1.19.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
requests==2.24.0
requests-mock==1.8.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
